requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Order matters to the pip dependency resolver, so sorting this file
# changes how packages are installed.  New dependencies should be
# added in alphabetical order, however, some dependencies may need to
# be installed in a specific order.
#
# PBR should always appear first

Django==1.10
Pillow>=4.1.1
djangorestframework>=3.6.2
MySQL-python>=1.2.5
django-oauth-toolkit>=0.12.0
uwsgi
lxml>=3.7.0
qrcode
pycrypto>=2.6.0
redis>=2.10.5
pyBarcode>=0.7.0

# 管理uwsgi进程
supervisor

# 允许跨域请求
django-cors-headers

